Reduce duplication for change equivalence tests
